requirements.txt
PySide6
pyaudio
librosa
numpy
live2d-py
PyOpenGL
toml
openai